+++ src/auth/css/auth.css
@@ -44,7 +44,7 @@
   background-color: rgba(0, 0, 0, 0.7);
 }
 .auth div img {
-  width: 60vh;
+  width: 90vh;
   border-top-right-radius: 20px;
   border-top-left-radius: 20px;
 }
