+++ src/home/Home.jsx
@@ -1,25 +1,18 @@
+
 import escuela1 from "../img/escuela1.jpg";
 import escuela2 from "../img/escuela2.jpg";
 import escuela3 from "../img/escuela3.jpg";
 import escuela4 from "../img/escuela4.jpg";
-import Logo from "../img/Logo.png";
-import login from "../img/usuario.png";
-import "../style/home.css";
-import "../style/navbar.css";
-import "../style/footer.css";
 import Login from "../auth/pages/Login";
-import { useState } from "react";
+import "../style/home.css";
+import "../style/footer.css";
+import NavBar from "../components/NavBar";
+import { useContext, useState } from "react";
+import AppContext from "../AppContext";
 
 function Index() {
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
-  const openModal = () => {
-    setIsModalOpen(true);
-  };
-
-  const closeModal = () => {
-    setIsModalOpen(false);
-  };
+  const {isModalOpen} = useContext(AppContext);
+  
   return (
     <>
       <div className="layout">
@@ -30,24 +23,8 @@ function Index() {
           <img src={escuela4} alt="" />
         </div>
       </div>
-      {isModalOpen ? <Login isOpen={isModalOpen} onClose={closeModal} /> : ""}
-      <header className="header">
-        <div className="left">
-          <a href="">
-            <img src={Logo} alt="Logo de la escuela sin fondo" />
-          </a>
-        </div>
-        <ul>
-          <li>Nosotros</li>
-          <li>Contacto</li>
-          <li>Tienda</li>
-        </ul>
-        <div className="right">
-          <button onClick={openModal}>
-            <img src={login} alt="Boton ingresar a su cuenta de usuario" />
-          </button>
-        </div>
-      </header>
+        {isModalOpen ? <Login /> : ""}
+      <NavBar />
       <footer>
         <ul>
           <li>Rolling © 2023</li>
