+++ src/routes/PrivateRouter.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
+import { Routes, Route, Navigate } from "react-router-dom";
 import Register from "../auth/pages/Register";
 import Admin from "../admin/pages/Admin";
