+++ src/routes/PublicRouter.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
-import Login from "../auth/pages/Login";
 import Home from "../home/Home";
+import Admin from "../admin/pages/Admin";
 
 function PublicRouter() {
   return (
@@ -9,6 +9,7 @@ function PublicRouter() {
       <Routes>
         <Route path="/" element={<Navigate to="/home" replace />} />
         <Route exact path="/home" element={<Home />} />
+        <Route exact path="/admin" element={<Admin />} />
         <Route path="*" element={<Navigate to="/" replace />} />
       </Routes>
     </BrowserRouter>
